api/src/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision
numpy
flask>=2.0.0
gunicorn
flask-cors
Pillow